fix rate type

+++ src/app/dto/get.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass
@@ -14,4 +15,4 @@ class GetRateDTO:
     id: int
     base_currency: str
     target_currency: str
-    rate: float
+    rate: Decimal
